rename GotRepository folder
