+++ app/src/main/java/es/npatarino/android/gotchallenge/di/AppComponent.java
@@ -6,7 +6,7 @@ import com.tonilopezmr.interactorexecutor.MainThread;
 import javax.inject.Singleton;
 
 import dagger.Component;
-import es.npatarino.android.gotchallenge.domain.GotHouseRepository.GotCharacterRepositoryImp;
+import es.npatarino.android.gotchallenge.domain.repository.GotCharacterRepositoryImp;
 import okhttp3.OkHttpClient;
 
 /**
