+++ app/src/main/java/es/npatarino/android/gotchallenge/di/components/HousesComponent.java
@@ -8,7 +8,7 @@ import es.npatarino.android.gotchallenge.di.AppComponent;
 import es.npatarino.android.gotchallenge.di.modules.ActivityModule;
 import es.npatarino.android.gotchallenge.di.modules.HousesModule;
 import es.npatarino.android.gotchallenge.domain.GoTHouse;
-import es.npatarino.android.gotchallenge.domain.GotHouseRepository.GotHouseRepositoryImp;
+import es.npatarino.android.gotchallenge.domain.repository.GotHouseRepositoryImp;
 import es.npatarino.android.gotchallenge.domain.interactor.common.GetListUseCase;
 import es.npatarino.android.gotchallenge.presenter.HouseListPresenter;
 import es.npatarino.android.gotchallenge.view.fragment.GoTHousesListFragment;
