+++ app/src/main/java/es/npatarino/android/gotchallenge/di/modules/HousesModule.java
@@ -9,8 +9,8 @@ import dagger.Module;
 import dagger.Provides;
 import es.npatarino.android.gotchallenge.di.Activity;
 import es.npatarino.android.gotchallenge.domain.GoTHouse;
-import es.npatarino.android.gotchallenge.domain.GotHouseRepository.GotCharacterRepositoryImp;
-import es.npatarino.android.gotchallenge.domain.GotHouseRepository.GotHouseRepositoryImp;
+import es.npatarino.android.gotchallenge.domain.repository.GotCharacterRepositoryImp;
+import es.npatarino.android.gotchallenge.domain.repository.GotHouseRepositoryImp;
 import es.npatarino.android.gotchallenge.domain.interactor.common.GetListUseCase;
 import es.npatarino.android.gotchallenge.domain.interactor.common.GetListUseCaseImp;
 import es.npatarino.android.gotchallenge.presenter.HouseListPresenter;
