+++ app/src/test/java/es/npatarino/android/gotchallenge/domain/GotHouseRepository/TestableGotCharacterRepository.java
@@ -1,6 +1,7 @@
 package es.npatarino.android.gotchallenge.domain.GotHouseRepository;
 
 import es.npatarino.android.gotchallenge.ResourceHelper;
+import es.npatarino.android.gotchallenge.domain.repository.GotCharacterRepositoryImp;
 
 /**
  * @author Antonio López.
